Address review comments
 - Split Jaeger tests out of test_integration
 - Use a non-default port. We now use a dynamic port
