+++ tests/system/apmserver.py
@@ -2,14 +2,13 @@ from datetime import datetime, timedelta
 import json
 import os
 import re
+import sets
 import shutil
-import threading
-import unittest
-from time import gmtime, strftime
-from urlparse import urlparse
-
 import sys
+import threading
 import time
+import unittest
+from urlparse import urlparse
 
 from elasticsearch import Elasticsearch
 from nose.tools import nottest
@@ -46,7 +45,7 @@ class BaseTest(TestCase):
     @classmethod
     def setUpClass(cls):
         cls.apm_version = "8.0.0"
-        cls.day = strftime("%Y.%m.%d", gmtime())
+        cls.day = time.strftime("%Y.%m.%d", time.gmtime())
         cls.beat_name = "apm-server"
         cls.beat_path = os.path.abspath(os.path.join(
             os.path.dirname(__file__), "..", ".."))
@@ -372,6 +371,66 @@ class ElasticTest(ServerBaseTest):
             if jline.get("logger") == "request" and u.path == url:
                 yield jline
 
+    def approve_docs(self, base_path, received, doc_type):
+        """
+        approve_docs compares the received documents to those contained
+        in the file at ${base_path}.approved.json. If that file does not
+        exist, then it is considered equivalent to a lack of documents.
+
+        Only the document _source is compared, and we ignore differences
+        in some context-sensitive fields such as the "observer", which
+        may vary between test runs.
+        """
+        base_path = self._beat_path_join(os.path.dirname(__file__), base_path)
+        approved_path = base_path + '.approved.json'
+        received_path = base_path + '.received.json'
+
+        try:
+            with open(approved_path) as f:
+                approved = json.load(f)
+        except IOError:
+            approved = []
+
+        received = [doc['_source'] for doc in received]
+        received.sort(key=lambda source: source[doc_type]['id'])
+
+        try:
+            for rec in received:
+                # Overwrite received observer values with the approved ones,
+                # in order to avoid noise in the 'approvals' diff if there are
+                # any other changes.
+                #
+                # We don't compare the observer values between received/approved,
+                # as they are dependent on the environment.
+                rec_id = rec[doc_type]['id']
+                rec_observer = rec['observer']
+                self.assertEqual(sets.Set(rec_observer.keys()), sets.Set(
+                    ["hostname", "version", "id", "ephemeral_id", "type", "version_major"]))
+                assert rec_observer["version"].startswith(str(rec_observer["version_major"]) + ".")
+                for appr in approved:
+                    if appr[doc_type]['id'] == rec_id:
+                        rec['observer'] = appr['observer']
+                        break
+            assert len(received) == len(approved)
+            for i, rec in enumerate(received):
+                appr = approved[i]
+                rec_id = rec[doc_type]['id']
+                assert rec_id == appr[doc_type]['id'], "New entry with id {}".format(rec_id)
+                for k, v in rec.items():
+                    self.assertEqual(v, appr[k])
+        except Exception as exc:
+            with open(received_path, 'w') as f:
+                json.dump(received, f, indent=4, separators=(',', ': '))
+
+            # Create a dynamic Exception subclass so we can fake its name to look like the original exception.
+            class ApprovalException(Exception):
+                def __init__(self, cause):
+                    super(ApprovalException, self).__init__(cause.message)
+
+                def __str__(self):
+                    return self.message + "\n\nReceived data differs from approved data. Run 'make update' and then 'approvals' to verify the diff."
+            ApprovalException.__name__ = type(exc).__name__
+            raise ApprovalException, exc, sys.exc_info()[2]
 
 class ClientSideBaseTest(ServerBaseTest):
     sourcemap_url = 'http://localhost:8200/assets/v1/sourcemaps'
