+++ tests/system/test_integration.py
@@ -9,7 +9,6 @@ import requests
 from apmserver import integration_test
 from apmserver import ClientSideElasticTest, ElasticTest, ExpvarBaseTest
 from apmserver import OverrideIndicesTest, OverrideIndicesFailureTest
-from sets import Set
 
 
 @integration_test
@@ -101,73 +100,6 @@ class Test(ElasticTest):
         self.approve_docs('error', rs['hits']['hits'], 'error')
 
         self.check_backend_error_sourcemap(self.index_error, count=4)
-
-    def test_jaeger_http(self):
-        """
-        This test sends a Jaeger span in Thrift encoding over HTTP, and verifies that it is indexed.
-        """
-        jaeger_span_thrift = self.get_testdata_path('jaeger', 'span.thrift')
-        self.load_docs_with_template(jaeger_span_thrift, self.jaeger_http_url, 'transaction', 1,
-                                     extra_headers={"content-type": "application/vnd.apache.thrift.binary"})
-        self.assert_no_logged_warnings()
-
-        # compare existing ES documents for errors with new ones
-        rs = self.es.search(index=self.index_transaction)
-        assert rs['hits']['total']['value'] == 1, "found {} documents".format(rs['count'])
-        self.approve_docs('jaeger_span', rs['hits']['hits'], 'transaction')
-
-    def approve_docs(self, base_path, received, doc_type):
-        base_path = self._beat_path_join(os.path.dirname(__file__), base_path)
-        approved_path = base_path + '.approved.json'
-        received_path = base_path + '.received.json'
-
-        try:
-            with open(approved_path) as f:
-                approved = json.load(f)
-        except IOError:
-            approved = []
-
-        received = [doc['_source'] for doc in received]
-        received.sort(key=lambda source: source[doc_type]['id'])
-
-        try:
-            for rec in received:
-                # Overwrite received observer values with the approved ones,
-                # in order to avoid noise in the 'approvals' diff if there are
-                # any other changes.
-                #
-                # We don't compare the observer values between received/approved,
-                # as they are dependent on the environment.
-                rec_id = rec[doc_type]['id']
-                rec_observer = rec['observer']
-                self.assertEqual(Set(rec_observer.keys()), Set(
-                    ["hostname", "version", "id", "ephemeral_id", "type", "version_major"]))
-                assert rec_observer["version"].startswith(str(rec_observer["version_major"]) + ".")
-                for appr in approved:
-                    if appr[doc_type]['id'] == rec_id:
-                        rec['observer'] = appr['observer']
-                        break
-            assert len(received) == len(approved)
-            for i, rec in enumerate(received):
-                appr = approved[i]
-                rec_id = rec[doc_type]['id']
-                assert rec_id == appr[doc_type]['id'], "New entry with id {}".format(rec_id)
-                for k, v in rec.items():
-                    self.assertEqual(v, appr[k])
-        except Exception as exc:
-            with open(received_path, 'w') as f:
-                json.dump(received, f, indent=4, separators=(',', ': '))
-
-            # Create a dynamic Exception subclass so we can fake its name to look like the original exception.
-            class ApprovalException(Exception):
-                def __init__(self, cause):
-                    super(ApprovalException, self).__init__(cause.message)
-
-                def __str__(self):
-                    return self.message + "\n\nReceived data differs from approved data. Run 'make update' and then 'approvals' to verify the diff."
-            ApprovalException.__name__ = type(exc).__name__
-
-            raise ApprovalException, exc, sys.exc_info()[2]
 
 
 @integration_test
